Main figure of estimated trends (wild and total spawners)

This is the main figure function. Not exported. It is used by NWCTrends_report and inst/doc/report_files/esu_report.Rmd.

The dots are the raw spawner counts, the black line is the smoothed total spawners estimate, and the red line is the smoothed wild spawners estimate which is "smoothed total estimate x smoothed fracwild estimate". Note that the wild spawner estimate is only shown from 1 year before and one year after the last actual fracwild estimate (in the data file). This is done so that the wild estimate does not over-extend the fracwild data. Fracwild estimates can be interpolated for missing years, but would not be appropriate to extend much before or past actual observed (or expert) fracwild data.

For the smoothed total estimates, information from all populations (via a non-diagonal year-to-year variance matrix) is used to estimate missing values and to account for observation error in the total spawner count. Because data from all populations are used, estimates can be made even for missing years at the beginning of the time series if there is data for those early years in other populations.

Status_trendfigure_multipanel(
  esu,
  pops,
  total.fit,
  fracwild.fit,
  plot.min.year = NULL,
  plot.max.year = NULL,
  silent = FALSE,
  CI.method = "hessian",
  CI.sim = 1000,
  log.scale = FALSE,
  same.scale = FALSE,
  nwctrends.options = NULL
)

Arguments

esu
The name of the ESU

pops
The population names that will be plotted (populations with too few data are eliminated)

total.fit
total fit returned by trend_fits()

fracwild.fit
fracwild fit returned by trend_fits()

plot.min.year
The x axis minimum.

plot.max.year
The x axis maximum.

silent
No output

CI.method
Method sent to MARSSparamCIs

CI.sim
If doing bootstrap CI, this is the number of bootstraps sent to MARSSparamCIs

log.scale
Put plot on log-scale versus the original raw scale

same.scale
Tweak the scale of wild and total in graph. Not used.

nwctrends.options
A list of plot options to change the appearance (colors, line types, line widths, point types, etc) in the plots. See nwctrends.options for a description of the options. Note, if risk_plot_multipanel() is called from NWCTrends_report() then nwctrends.options has already been set and can be left at NULL in this call.

Value

A plot

See also

Author

Eli Holmes, NOAA, Seattle, USA. eli(dot)holmes(at)noaa(dot)gov