Productivity plot

This uses the smoothed spawner estimates and smoothed fracwild estimates to compute a productivity metric. Type 3: wild(t+1)/wild(t). Type 1: wild(t+lag)/total(t), where wild is smoothed total estimate times smoothed fracwild estimate and total is the smoothed total estimate.

In the Viability Report, type=1 and the lag is set to 3 or 4 (depending on species).

productivity_plot(
  esu,
  pops,
  total.fit,
  fracwild.fit,
  min.year = NULL,
  max.year = NULL,
  type = 1,
  lag = 4,
  nwctrends.options = NULL
)

Arguments

esu

The name of the ESU

pops

The population names that will be plotted.

total.fit

total fit returned by trend_fits()

fracwild.fit

fracwild fit returned by trend_fits()

min.year

The x axis minimum. First year for numerator.

max.year

The x axis maximum. Last year for numerator.

type

The type of plot. Type 3: wild(t+1)/wild(t). Type 1: wild(t+lag)/total(t)

lag

The number of years prior to use in the denominator, e.g. spawnwers(year-lag). Note not used if type=3.

nwctrends.options

A list of plot options to change the appearance (colors, line types, line widths, point types, etc) in the plots. See nwctrends.options for a description of the options. Note, if risk_plot_multipanel() is called from NWCTrends_report() then nwctrends.options has already been set and can be left at NULL in this call.

Value

A plot

Author

Eli Holmes, NOAA, Seattle, USA. eli(dot)holmes(at)noaa(dot)gov