Create the trend tables

Create the tables with the trends for different time periods using the smoothed spawner estimates. The trend is the slope of a linear regression of the log spawner counts versus year. Despite the name of the function, the range of years need not be 15 years. The years to show are specified by list year.ranges. The ranges are specified as begin.year:end.year, for example 1990:2005. year.ranges can be padded into the NWCTrends_report() call by passing in trend.table.control as list. For example list(year.ranges=list(1990:2000,2000:2010)).

trend_15_table(
  pops,
  mpg,
  total.fit,
  fracwild.fit,
  year.ranges = list(1990:2005, 1999:2014),
  wild = TRUE,
  navalue = " "
)

Arguments

pops

which populations to include in the table

mpg

Population group. Shown in the table.

total.fit

The matrix of total spawner estimates

fracwild.fit

The matrix of fraction wild associated with each total row.

year.ranges

The columns of years.

wild

Show smoothed wild or smoothed total.

navalue

Value to use for NAs in the table. Default is a blank.

Value

A data frames with the estimates trend for each year range in a different column.

Author

Eli Holmes, NOAA, Seattle, USA. eli(dot)holmes(at)noaa(dot)gov